feature: modifyStudy 유효성검사 추가
