+++ client/src/pages/ModifyUser.js
@@ -1,0 +1,5 @@
+function ModifyUser() {
+  return <div>ModifyUser</div>;
+}
+
+export default ModifyUser;
